update admin backfill, more server cleanup
