+++ apps/web/src/lib/types/profile.ts
@@ -1,0 +1,6 @@
+export type Profile = {
+	did: string;
+	handle: string;
+	displayName: string | null;
+	avatar: string | null;
+};
